Explicit casts for status (int <-> uint32_t)
proto_dynamodb_kv_client.c:
    status is exactly 1 or 2, so it's safe to cast to an (int).

lib/proto_dynamodb_kv/proto_dynamodb_kv_server.c:
lib/proto_kvlds/proto_kvlds_server.c:
    The sanity checks ensure that status matches the allowable values
    (according to the header comment and/or INTERFACES).

+++ lib/proto_dynamodb_kv/proto_dynamodb_kv_client.c
@@ -92,7 +92,7 @@ callback_data(void * cookie, uint8_t * buf, size_t buflen)
 		case 2:
 			if (buflen != 4)
 				BAD("GET", "bogus length");
-			failed = status;
+			failed = (int)status;
 			goto failed;
 		default:
 			BAD("GET", "invalid status");
